requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
Django==4.0.5
django-slack==5.17.8
djangorestframework==3.13.1
Flask==2.1.2
frozenlist==1.3.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.2
pyee==9.0.4
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
slackclient==2.9.4
slacker==0.9.65
slacker-log-handler==1.8.0
slackeventsapi==3.0.1
sqlparse==0.4.2
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
Werkzeug==2.1.2
yarl==1.7.2
